Refactor core logics' dir for v2
