+++ src/components/v1/molecules/CollegeCalendarBase.tsx
@@ -2,7 +2,7 @@ import React from "react";
 import { CalendarBase } from "@/components/v1/atoms/CalendarBase";
 import { useResourceGroupLabelContentInTemplateCalendar } from "@/hooks/v1/useResourceGroupLabelContentInTemplateCalendar";
 import { useStory } from "@/hooks/v1/useStory";
-import { BaseStory } from "@/core/story/BaseStory";
+import { BaseStory } from "@/core/v1/story/BaseStory";
 import { OPEN_OPTION_BUTTON } from "@/hooks/v1/useTemplateCustomButtons";
 
 const configs = {
