+++ src/components/v1/molecules/EventEditModal.tsx
@@ -10,7 +10,7 @@ import { useForm } from "react-hook-form";
 import { useSelector } from "react-redux";
 import { selectEventModal, selectEvent } from "@/redux/v1/ui/eventModal";
 import { useEvent } from "@/hooks/v1/useEvent";
-import { BaseEvent } from "@/core/event/BaseEvent";
+import { BaseEvent } from "@/core/v1/event/BaseEvent";
 
 interface Props {
   isOpen: boolean;
