--- src/core/calendar/UserCalendar/model.ts
@@ -1,3 +1,0 @@
-import { BaseCalendar } from "@/core/calendar/BaseCalendar";
-
-export type UserCalendar = BaseCalendar;
