--- src/core/story/PrivateCollegeStory/model.ts
@@ -1,3 +1,0 @@
-import { BaseStory } from "@/core/story/BaseStory";
-
-export type PrivateCollegeStory = BaseStory;
