+++ src/core/v1/calendar/UserCalendar/createUserCalendar.ts
@@ -1,5 +1,5 @@
-import { UserCalendar } from "@/core/calendar/UserCalendar/model";
-import { ProfileStory } from "@/core/story/ProfileStory/model";
+import { UserCalendar } from "@/core/v1/calendar/UserCalendar/model";
+import { ProfileStory } from "@/core/v1/story/ProfileStory/model";
 
 type Story = ProfileStory;
