+++ src/core/v1/calendar/UserCalendar/model.ts
@@ -1,0 +1,3 @@
+import { BaseCalendar } from "@/core/v1/calendar/BaseCalendar";
+
+export type UserCalendar = BaseCalendar;
