+++ src/core/v1/event/createProfileEvents.ts
@@ -8,7 +8,7 @@ import { getRangeNumbers } from "@/lib/age";
 import { uuid } from "@/lib/uuid";
 import { createDate } from "@/lib/date";
 import { WORKING_HOLIDAY_APPLICATION_LIMITATION_AGE } from "@/constants/visa";
-import { BaseEvent } from "@/core/event/BaseEvent";
+import { BaseEvent } from "@/core/v1/event/BaseEvent";
 
 type CreateProfileEventsParams = {
   startDate: Date;
