+++ src/core/v1/resource/BaseResource.ts
