+++ src/core/v1/resource/createProfileResources.ts
