+++ src/core/v1/story/PrivateCollegeStory/model.ts
@@ -1,0 +1,3 @@
+import { BaseStory } from "@/core/v1/story/BaseStory";
+
+export type PrivateCollegeStory = BaseStory;
