+++ src/hooks/v1/usePrivateCollegeCalendar.tsx
@@ -1,9 +1,9 @@
 import React from "react";
 import { useDispatch, useSelector } from "react-redux";
-import { BaseEvent } from "@/core/event/BaseEvent";
-import { BaseResource } from "@/core/resource/BaseResource";
-import { TemplateOption } from "@/core/calendar/BaseCalendar";
-import { createPrivateCollegeCalendar } from "@/core/calendar/TemplateCalendar/createCalendar";
+import { BaseEvent } from "@/core/v1/event/BaseEvent";
+import { BaseResource } from "@/core/v1/resource/BaseResource";
+import { TemplateOption } from "@/core/v1/calendar/BaseCalendar";
+import { createPrivateCollegeCalendar } from "@/core/v1/calendar/TemplateCalendar/createCalendar";
 import { createDate } from "@/lib/date";
 import {
   upsertPrivateCollegeCalendarAction,
