+++ src/hooks/v1/useResourceGroupLabelContentInTemplateCalendar.tsx
@@ -5,7 +5,7 @@ import {
   selectPublicCollegeCalendar,
 } from "@/redux/v1/features/templateCalendar";
 import { selectUserCalendar } from "@/redux/v1/features/userCalendars";
-import { updateStory } from "@/core/story/BaseStory";
+import { updateStory } from "@/core/v1/story/BaseStory";
 import styled from "styled-components";
 import Button from "@mui/material/Button";
