+++ src/hooks/v1/useUserCalendar.tsx
@@ -2,9 +2,9 @@ import React from "react";
 import { useDispatch, useSelector } from "react-redux";
 import { DateSelectArg } from "@fullcalendar/react";
 import { uuid } from "@/lib/uuid";
-import { createProfileStory } from "@/core/story/ProfileStory/createProfileStory";
-import { initEvent } from "@/core/event/BaseEvent";
-import { createUserCalendar } from "@/core/calendar/UserCalendar/createUserCalendar";
+import { createProfileStory } from "@/core/v1/story/ProfileStory/createProfileStory";
+import { initEvent } from "@/core/v1/event/BaseEvent";
+import { createUserCalendar } from "@/core/v1/calendar/UserCalendar/createUserCalendar";
 import {
   updateCalendarsAction,
   removeCalendarAction,
