+++ src/redux/v1/features/templateCalendar.test.ts
@@ -9,7 +9,7 @@ import reducer, {
 import {
   PRIVATE_COLLEGE_CALENDAR_ID,
   PUBLIC_COLLEGE_CALENDAR_ID,
-} from "@/core/calendar/TemplateCalendar/model";
+} from "@/core/v1/calendar/TemplateCalendar/model";
 import { RootState } from "@/redux/rootReducer";
 import {
   privateCollegeCalendarFactory,
