+++ src/redux/v1/features/templateCalendar.ts
@@ -3,11 +3,11 @@ import { RootState } from "@/redux/rootReducer";
 import {
   PRIVATE_COLLEGE_CALENDAR_ID,
   PUBLIC_COLLEGE_CALENDAR_ID,
-} from "@/core/calendar/TemplateCalendar/model";
+} from "@/core/v1/calendar/TemplateCalendar/model";
 import {
   PrivateCollegeCalendar,
   PublicCollegeCalendar,
-} from "@/core/calendar/TemplateCalendar/model";
+} from "@/core/v1/calendar/TemplateCalendar/model";
 
 type UpsertPrivateCollegeCalendarPayload = {
   calendar: PrivateCollegeCalendar;
